==> 03 Scripts/1_Data_Import_basic_analysis.ipynb <==
# 1 Importing data and basic statistical analysis

Contents

1. Import libraries
2. Practice datatypes

1. Import libraries

#import libraries
import pandas as pd
import numpy as np
import os

2. Practice datatypes

#Code 3 different ways of reaching a result of 100 by adding or subtracting numeric variables.
a = 150
b = 50
print(a-b)

print(b + b)

c = 25
print(b + c +c)

#Code 2 floating-point variables and divide them by each other.
d = 3.5
e = 1.1
print(d / e)

#Construct a short word made of separate strings 
f = 'Insta'
g = 'cart'
print(f + g)

#Construct 2 short sentences made of separate strings
m = 'My '
k = 'cat '
i = 'is '
n = 'snoring'
o = 'Other '
p = 'sleeping'
q = '.'

print(m + k + i + n + q)
print(o + k + i + p + q)

#Bonus task At the bottom of your file, include a command that makes a string appear 100 times.
print(k *100)
