==> 03 Scripts/2_Data_Checks_OrdersData_ProductsData.ipynb <==
# Data Checks on Orders and Products datasets

Contents

1. Import libraries
2. Read, transform, and describe orders data
3. Read, transform, and describe products data

1. Import libraries

#import libraries
import pandas as pd
import numpy as np
import os

2. Read, transform, and describe orders data

#Assign project folder path to string variable 'path'
path = r'C:\Users\Caroline\Documents\Career Foundry\03 2023 Instacart Basket Analysis'

# Assign columns to select to list 'vars_list'
vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

#Import orders.csv data set using os library and vars_list shortcut, omitting 'eval_set' column
df = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), usecols = vars_list)

#Print the first 5 rows of your dataframe.
print(df.head(5))

3. Read, transform, and describe products data

# Import products.csv as df_prods
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

# Print the first 20 rows and last 35 rows of the df_prods dataframe
print(df_prods.head(20))
print(df_prods.tail(35))

# Print the names of the columns in the df_prods dataframe
print(df_prods.columns)

# Print the number of rows and columns in the df_prods dataframe
print(df_prods.shape)

11a - maximum value of aisle_id = 49693
11b - department_id data type is integer (64 bit integer).

# Use .info() to get summary information on df_prods to answer #11
df_prods.info()
