==> 03 Scripts/3_Data_wrangling.ipynb <==
# 3 Data Wrangling and Subsetting

Contents

1. Importing libraries
2. Importing data
3. Data Wrangling - Orders Dataframe
4. Data Wrangling - Departments Dataframe
    4.1 Creating data dictionary
5. Data Wrangling - Products Dataframe
6. Task 4.4 Questions
7. Export dataframes


1. Importing libraries

#Import libraries
import pandas as pd
import numpy as np
import os

2. Importing data

path = r'C:\Users\Caroline\Documents\Career Foundry\03 2023 Instacart Basket Analysis'

#import orders.csv as df_ords
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), index_col = False)

#import products.csv as df_prods
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

#import departments.csv as df_dep
df_dep = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'departments.csv'), index_col = False)

3. Data Wrangling - Orders Dataframe

# dropping the 'eval_set' column from df_ords and overwriting df_ords
df_ords = df_ords.drop(columns = ['eval_set'])

df_ords.head()

#checking for missing values
df_ords['days_since_prior_order'].value_counts(dropna = False)
# Result - 206209 missing values (NaN). Column is still usable since null are not the majority of the column

# rename the orders_dow column to orders_day_of_week
df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

# change the data type of order_id to a string since it does not need to be an integer
df_ords['order_id'] = df_ords['order_id'].astype('str')

# inspect df_ords
df_ords.info()

df_ords.describe()

# change 'orders_day_of_week', 'order_hour_of_day', and 'order_number' to int8, change 'user_id' to string
df_ords['orders_day_of_week'] = df_ords['orders_day_of_week'].astype('int8')
df_ords['order_hour_of_day'] = df_ords['order_hour_of_day'].astype('int8')
df_ords['user_id'] = df_ords['user_id'].astype('str')
df_ords['order_number'] = df_ords['order_number'].astype('int8')

# change 'days_since_prior_order' to float16
df_ords['days_since_prior_order'] = df_ords['days_since_prior_order'].astype('float16')

# inspect df_ords
df_ords.info()

4. Data Wrangling - Departments Dataframe

# transpose df_dep
df_dep_t = df_dep.T

df_dep_t.head()

# add index to df_dep_t
df_dep_t.reset_index()

# Create a new header by defining the first row to a variable
new_header = df_dep_t.iloc[0]

# Delete first row of df_dep_t by copying all of the list except for the first row
df_dep_t_new = df_dep_t[1:]

# Set new_header variable as the new header for the df_dep_t_new data frame
df_dep_t_new.columns = new_header

# Check df_dep_t_new
df_dep_t_new

4.1  Creating data dictionary

# make df_dep_t_new into data dictionary
data_dict = df_dep_t_new.to_dict('index')

# inspect df_dep_t_new
df_dep_t_new.info()

5. Data Wrangling - Products Dataframe

# inspecting df_prods
df_prods.info()

df_prods.describe()

# change 'aisle_id' and 'department_id' to int8, change 'product_id' to int32
df_prods['aisle_id'] = df_prods['aisle_id'].astype('int8')
df_prods['department_id'] = df_prods['department_id'].astype('int8')
df_prods['product_id'] = df_prods['product_id'].astype('int32')

# change 'prices' to float32
df_prods['prices'] = df_prods['prices'].astype('float32')

# inspecting df_prods
df_prods.info()

#6 - Task 4.4 

# Find another identifier variable in the df_ords dataframe that doesn't need to be a numeric variable
#Copied from above - df_ords['user_id'] = df_ords['user_id'].astype('str')

# Look for a variable in your df_ords dataframe with an unintuitive name and change its name without overwriting the data frame
df_ords.columns

df_ords.rename(columns = {'orders_day_of_week' : 'order_day_of_week'})

# Make a frequency count of 'order_hour_of_day'
df_ords['order_hour_of_day'].value_counts()

#4 The busiest hours for orders placed are 10, 11, 15, 14, and 13.

# Det. the meaning behind a value of 4 in the "department_id" column within the df_prods dataframe using a data dictionary
print(data_dict.get('4'))

#5 Answer - the value 4 in df_dep_t_new 'department_id' column is the produce department

# create breakfast subset
df_breakfast = df_prods[df_prods['department_id']==14]

df_breakfast.head()

# create dinner party subset
df_dinner_party = df_prods.loc[df_prods['department_id'].isin([5, 7, 12, 20])]

df_dinner_party.head()

# How many rows does the df_dinner_party have?
df_dinner_party.shape

#8 df_dinner_party has 7650 rows.

# extract infomation of user_id = 1

df_ords_user1 = df_ords[df_ords['user_id'] == '1' ]

df_ords_user1

# Summarize data for user_id 1
df_ords_user1.describe()

7. Export dataframes

#Export df_ords 
df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_wrangled.csv'))

#Export df_dep_t_new
df_dep_t_new.to_csv(os.path.join(path, '02 Data','Prepared Data', 'departments_wrangled.csv'))

#Export df_prods
df_prods.to_csv(os.path.join(path, '02 Data','Prepared Data', 'products_wrangled.csv'))

#Export df_ords without index? to fix 4.5 problem
df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_wrangled2.csv'), index_col = False)

#Export df_dep_t_new index? to fix 4.5 problem
df_dep_t_new.to_csv(os.path.join(path, '02 Data','Prepared Data', 'departments_wrangled2.csv'), index_col = False)

#Export df_prods index? to fix 4.5 problem
df_prods.to_csv(os.path.join(path, '02 Data','Prepared Data', 'products_wrangled2.csv'), index_col = False)
