==> 03 Scripts/4_Data_Consistency_Checks.ipynb <==
# 4 Data Consistency Checks

Contents

1. Importing libraries
2. Import data
3. Checking Products dataframe
    3.1 Missing values and duplicates
    3.2 Testing for mixed datatypes
    3.3 Correcting error in aisle_id
    3.4 Final data checks
4. Checking Orders dataframe
    4.1 Testing for mixed datatypes
    4.2 Missing values and duplicates
5. Exporting data

1. Importing libraries

#Import libraries
import pandas as pd
import numpy as np
import os

path = r'C:\Users\Caroline\Documents\Career Foundry\03 2023 Instacart Basket Analysis'

2. Import Data

#import orders_wrangled.csv as df_ords
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_wrangled.csv'), index_col = 0)

#import products_wrangled.csv as df_prods
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'products_wrangled.csv'), index_col = 0)

3. Checking Products dataframe

3.1 Missing values and duplicates

#  to find missing values
df_prods.isnull().sum()

# create a dataframe with nulls in the product_name column
df_nan = df_prods[df_prods['product_name'].isnull() == True]

df_nan

# Create a dataframe where all the missing product names are not included 
df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

# check for full duplicates
df_dups = df_prods_clean[df_prods_clean.duplicated()]

# Which values are duplicated?
df_dups

# check the current number of rows in df_prods_clean
df_prods_clean.shape

# new data frome without duplicates
df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

# check shape of products dataframe after removing duplicates
df_prods_clean_no_dups.shape

3.2 Testing for mixed datatypes

# Search for mixed data types in df_prods_clean_no_dups
for col in df_prods_clean_no_dups.columns.tolist():
  weird = (df_prods_clean_no_dups[[col]].applymap(type) != df_prods_clean_no_dups[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_prods_clean_no_dups[weird]) > 0:
    print (col)
print ('test complete - this statement will always print')

3.3 Correcting error in aisle_id

#Descriptive statistics of df_prods_clean_no_dups
df_prods_clean_no_dups.describe()

A negative value for aisle_id is odd, and the max value in prices is unusual. 

# Search for records with negative aisle_is value
df_prods_clean_no_dups.loc[df_prods_clean_no_dups['aisle_id'] < 0]

#import original products.csv as df_prods_ori
df_prods_ori = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

# Search for records with negative aisle_id value
df_prods_ori.loc[df_prods_ori['aisle_id'] < 0]

After doing research, changing the data type to int8 in Task 4.4 turned values that exceeded 127 to negative values. Aisle_id had values to 134. Solution - overwrite the aisle_id from original data after repeating missing values & duplicate value steps so that the indexes of both dataframes match.

# Create a dataframe where all the missing product names are not included 
df_prods_clean2 = df_prods_ori[df_prods_ori['product_name'].isnull() == False]

# new data frome without duplicates
df_prods_clean2_no_dups = df_prods_clean2.drop_duplicates()

# check shape of products dataframe after removing duplicates
df_prods_clean2_no_dups.shape

#overwrite aisle_id column on df_prods_clean_no_dups
df_prods_clean_no_dups.eval('aisle_id = @df_prods_clean2_no_dups.aisle_id', inplace = True)

3.4 Final data checks

# Re-check the descriptive statistics of df_prods_clean_no_dups
df_prods_clean_no_dups.describe()

# Search for records with prices over 40
df_prods_clean_no_dups.loc[df_prods_clean_no_dups['prices'] > 40]

#  Find median value of column 'prices'
df_prods_clean_no_dups['prices'].median()

# Replace 'prices' for product_id 21533 and 33666 with the median value 
df_prods_clean_no_dups = df_prods_clean_no_dups.replace ({"prices":{99999.0: 7.1, 14900.0:7.1 }})

# check that changes were sucessful
df_prods_clean_no_dups[df_prods_clean_no_dups["product_id"]==21553]

df_prods_clean_no_dups[df_prods_clean_no_dups["product_id"]==33664]

# Re-check the descriptive statistics of df_prods_clean_no_dups
df_prods_clean_no_dups.describe()

Descriptive data looks good. Max price of 25 is reasonable. No negative values.

#Export df_prod_clean_no_dups 
df_prods_clean_no_dups.to_csv(os.path.join(path, '02 Data','Prepared Data', 'products_checked.csv'))

# Checking Orders DF for missing values and duplications 

4. Checking Orders dataframe

# describe the Orders dataframe
df_ords.describe()

# Check df_ords size
df_ords.shape

2. The 'days_since_prior_order_ column has a count fewer then the other column. Possible missing vales that might need to be flagged or dropped.
Min values are all either 1 or 0. Max order_id is the same as the count. Fewer user_id then the count makes sense if users have multiple orders. 0-7 for orders_day_of_the_week for 7 total makes sense. and 0-23 for order_hour_of_day. The maximum value of days_since_prior_order is 30.

2 - Correction - O-6 is the correct range for orders_day_of_the_week. Since the count starts at 0, there are 7 total values.

4.1 Testing for mixed datatypes

# Search for mixed data type in df_ords
for col in df_ords.columns.tolist():
  weird = (df_ords[[col]].applymap(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird]) > 0:
    print (col)
print ('test complete - this statement will always print')

#4 If you find mixed-type data, fix it. The column in question should contain observations of a single data type
No changes made.

4.2 Check for missing values and duplicates

# Run a check for missing values in df_ords
#  to find missing values
df_ords.isnull().sum()

#5a There are 206,209 null values in the 'days since prior order' column. This is likely to be customers who have only used Instacart once

#6 Address the missing values using an appropriate method
# Create subset to show missing values in df_ords
df_ords_missing = df_ords[df_ords['days_since_prior_order'].isnull() == True]

# View missing values subset
df_ords_missing

# If the max value of 'order_number' is 1, then all of the missing values in days since prior orders are from customers 
# who have only used the servuce once
df_ords_missing['order_number'].max()

#6 It does not make sense to impute values in this situation or to remove the data. One option is to add a column to indicate a repeat or non-repeat customer. However a value of 1 in the 'order_number' column can also as a flag for a new customer so I have chosen not to make any changes yet.

# 7. Run a check for duplicate values in your df_ords data
df_dups_ords = df_ords[df_ords.duplicated()]

df_dups_ords

#7a There are no full duplicates. No changes to be made.

5. Exporting data

#Export df_prod_clean_no_dups 
df_prods_clean_no_dups.to_csv(os.path.join(path, '02 Data','Prepared Data', 'products_checked.csv'))

#Export df_ords 
df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_checked.csv'))
