==> 03 Scripts/5_Merging_Datasets.ipynb <==
# 5 Merging Datasets

Contents

1. Importing Libraries
2. Importing Data
    2.1 Data Wrangling & Consistency checks on df_ords_prior
3. Checking and Merging Dataframes 
4. Export Merged Dataframe

#01 Importing Libraries

#Import libraries
import pandas as pd
import numpy as np
import os

#02 Importing Data

#Folder Path
path = r'C:\Users\Caroline\Documents\Career Foundry\03 2023 Instacart Basket Analysis'

#import orders_checked.csv as df_ords
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_checked.csv'), index_col = 0)

#import order_products_prior.csv as df_ords_prior
df_ords_prior = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'order_products__prior.csv'), index_col = False)

#02.1 Data Wrangling & Consistency checks on df_ords_prior

#  Missing values check
df_ords_prior.isnull().sum()

# Checking for full duplicates
df_dups = df_ords_prior[df_ords_prior.duplicated()]

df_dups

# Checking for mixed data types in df_ords_prior
for col in df_ords_prior.columns.tolist():
  weird = (df_ords_prior[[col]].applymap(type) != df_ords_prior[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords_prior[weird]) > 0:
    print (col)
print ('test complete - this statement will always print')

#Descriptive statistics of df_ords_prior
df_ords_prior.describe()

# Confirming the shape of in df_ords_prior
df_ords_prior.shape

Data looks consistent and without null values, duplicates, and mixed data types. The 'add_to_cart_order' column could possibly be dropped but will be kept for now.

#03 Checking and Merging Dataframes

#check first rows of df_ords_prior
df_ords_prior.head()

#check first rows of df_ords
df_ords.head()

#Check df_ords_prior dataframe size
df_ords_prior.shape

# Check df_ords dataframe size
df_ords.shape

# Merge df_ords and df_ords_prior. Indicator = true to check if data is present in both dataframes
df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)

#Check output of merged dataframes
df_merged_large.head()

# Check value count for '_merge' column
df_merged_large['_merge'].value_counts()

#Check df_merged_large for file size and number of rows
df_merged_large.info()

#04 Export Merged Dataframe

#Export df_merged_large in .pkl format
df_merged_large.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_combined.pkl'))
