==> 03 Scripts/6_Combining_Exporting_data.ipynb <==
# 6 Combining and Exporting Data

Contents

1. Importing Libraries
2. Importing Data
3. Checking Dataframes
4. Merging Dataframes
5. Export Data

#01 Importing Libraries

#Import libraries
import pandas as pd
import numpy as np
import os

#02 Importing Data

#Folder Path
path = r'C:\Users\Caroline\Documents\Career Foundry\03 2023 Instacart Basket Analysis'

#import products_checked.csv as df_ords
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'products_checked.csv'), index_col = 0)

#import orders_products_combined.pkl as df_merged_orders
df_merged_orders = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_combined.pkl'))

#03 Checking Dataframes

#Check df_prods dataframe size
df_prods.shape

#Check df_merged_orders dataframe size
df_merged_orders.shape

#check first rows of df_prods
df_prods.head()

#check first rows of df_merged_orders
df_merged_orders.head()

# Drop _merge column from df_merged_orders
df_merged_orders = df_merged_orders.drop(columns = ['_merge'])

#check first rows of df_merged_orders
df_merged_orders.head()

#04 Merging Dataframes

# Merge df_merged_orders and df_prods. Indicator = true to check if data is present in both dataframes
df_orders_products_merged = df_merged_orders.merge(df_prods, on = 'product_id', indicator = True)

# Check the output of merged dataset
df_orders_products_merged.head()

# Check value count for '_merge' column
df_orders_products_merged['_merge'].value_counts()

#Check size of df_orders_products_merged  
df_orders_products_merged.shape

#check data types of df_orders_products_merged
df_orders_products_merged.info()

#Change data types to reduce memory requirements
df_orders_products_merged['user_id'] = df_orders_products_merged['user_id'].astype('str')
df_orders_products_merged['order_number'] = df_orders_products_merged['order_number'].astype('int8')
df_orders_products_merged['orders_day_of_week'] = df_orders_products_merged['orders_day_of_week'].astype('int8')
df_orders_products_merged['order_hour_of_day'] = df_orders_products_merged['order_hour_of_day'].astype('int8')

#Change data types to reduce memory requirements
df_orders_products_merged['product_id'] = df_orders_products_merged['product_id'].astype('int32')
df_orders_products_merged['reordered'] = df_orders_products_merged['reordered'].astype('int8')
df_orders_products_merged['aisle_id'] = df_orders_products_merged['aisle_id'].astype('int16')
df_orders_products_merged['department_id'] = df_orders_products_merged['department_id'].astype('int8')

#Change data types to reduce memory requirements
df_orders_products_merged['days_since_prior_order'] = df_orders_products_merged['days_since_prior_order'].astype('float16')
df_orders_products_merged['prices'] = df_orders_products_merged['prices'].astype('float32')
df_orders_products_merged['add_to_cart_order'] = df_orders_products_merged['add_to_cart_order'].astype('int16')
df_orders_products_merged['order_id'] = df_orders_products_merged['order_id'].astype('int32')

# Drop _merge column from df_orders_products_merged
df_orders_products_merged = df_orders_products_merged.drop(columns = ['_merge'])

#check data types of df_orders_products_merged
df_orders_products_merged.info()

df_orders_products_merged.describe()

#05 Export Data

#Export df_orders_products_merged in .pkl format
df_orders_products_merged.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged.pkl'))
