==> 03 Scripts/7_Deriving_New_Variables.ipynb <==
# 7 Deriving New Variables

Contents

1. Importing Libraries
2. Importing Data
3. Create 'Price_label' function
4. Create 'Busiest Day' function
5. Create 'Busiest days' function
6. Determine 'Busiest Period of the Day'
7. Export Data

#01 Importing Libraries

#Import libraries
import pandas as pd
import numpy as np
import os

#02 Importing Data

#Folder Path
path = r'C:\Users\Caroline\Documents\Career Foundry\03 2023 Instacart Basket Analysis'

#import orders_products_merged.pkl as df_orders_products_merged
df_ords_prods_merge = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged.pkl'))

#check dataframe after importing
df_ords_prods_merge.shape

#03 Create 'Price_label function

#use loc() function to search the entire data frame
df_ords_prods_merge.loc[df_ords_prods_merge['prices'] > 15, 'price_range_loc'] = 'High range product'

# Mid range statement
df_ords_prods_merge.loc[(df_ords_prods_merge['prices'] <= 15) & (df_ords_prods_merge['prices'] > 5), 'price_range_loc'] = 'Mid range product'

# Low range loc() statement
df_ords_prods_merge.loc[df_ords_prods_merge['prices'] <= 5, 'price_range_loc'] = 'Low range product'

# to check the results of price_range function
df_ords_prods_merge['price_range_loc'].value_counts(dropna = False)

#04 Create 'Busiest Day' function

# Get a count of the orders placed on each day of the week
df_ords_prods_merge['orders_day_of_week'].value_counts(dropna = False)

# for loop - value is a place holder
result = []

for value in df_ords_prods_merge['orders_day_of_week']:
    if value == 0:
        result.append("Busiest day")
    elif value == 4:
         result.append("Least busy")
    else:
        result.append("Regularly busy")

#create a column in df_ords_prods_merge to place result
df_ords_prods_merge['busiest_day'] = result

# Count of busiest day
df_ords_prods_merge['busiest_day'].value_counts(dropna = False)

#05 Create 'Busiest days' function

# for loop to determine the 2 busiest (day 0 and 1) and 2 slowest days (day 4 and 3)
result2 = []

for value in df_ords_prods_merge['orders_day_of_week']:
    if value == 0 :
        result2.append("Busiest days")
    elif value == 1:
        result2.append("Busiest days")
    elif value == 4:
         result2.append("Slowest days")
    elif value == 3:
        result2.append("Slowest days")
    else:
        result2.append("Regularly busy")

#create a column in df_ords_prods_merge to place result
df_ords_prods_merge['busiest_period'] = result2

# Count of orders in the busiest period of the week
df_ords_prods_merge['busiest_period'].value_counts(dropna = False)

Observations : the top 2 busiest days are roughly only 1 millon fewer orders then the 3 average days combined.  

#06 Determine Busiest Period of the Day

# Determine the hours of the day with 'Most orders', 'Average orders', 'Fewest orders'
# Get a count of the orders placed on each hour of the day
df_ords_prods_merge['order_hour_of_day'].value_counts(dropna = False)

# values over 2,500,000 = "Most orders" , values less then 250,000  = 'Fewest orders', rest 'Average orders'
result3 = []

for value in df_ords_prods_merge['order_hour_of_day']:
    if (value >= 10) & (value <= 16):
        result3.append("Most orders")
    elif value < 6:
        result3.append("Fewest orders")
    else:
        result3.append("Average orders")

#create a new column in df_ords_prods_merge containing labels called "busiest_period_of_the_day" 
df_ords_prods_merge['busiest_period_of_the_day'] = result3

# Print the frequency for the busiest_period_of_the_day
df_ords_prods_merge['busiest_period_of_the_day'].value_counts(dropna = False)

#07 Export Data

df_ords_prods_merge.shape

#check data types of df_ords_prods_merge
df_ords_prods_merge.info()

df_ords_prods_merge.describe()

#Export df_orders_products_merged in .pkl format
df_ords_prods_merge.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_PR_busy.pkl'))
