==> 03 Scripts/8_Grouping_Aggregating_Data.ipynb <==
# 8 Grouping Data & Aggregating Variables

Contents

1. Importing Libraries
2. Importing Data
3. Determine aggregated mean of orders grouped by department
4. Create a loyalty flag for existing customers
5. Create 'loyalty_flag' column
6. Analyze spending habits of customers
7. Create a spending flag based on average price by user
8. Create a frequent customer flag
9. Export Data

#01 Importing Libraries

#Import libraries
import pandas as pd
import numpy as np
import os

#02 Importing Data

#Folder Path
path = r'C:\Users\Caroline\Documents\Career Foundry\03 2023 Instacart Basket Analysis'

#import orders_products_merged.pkl as df_orders_products_merged
ords_prods_merge = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_PR_busy.pkl'))

#check the imported data
ords_prods_merge.shape

ords_prods_merge.head()

#03 Determine aggregated mean of orders grouped by department

#split the data into groups based on department_id and use agg function to obtain mean values for the 'order_number column'
ords_prods_merge.groupby('department_id').agg({'order_number': ['mean']})

#04 Create a loyalty flag for existing customers

# First step - create a new column called 'max_order', then group ords_prods_merge by 'user_id', and transform() function 
# applied on the 'order_number ' column with np.max arguement
ords_prods_merge['max_order'] = ords_prods_merge.groupby(['user_id'])['order_number'].transform(np.max)

#check the new column with .head()
ords_prods_merge.head(15)

#to show all the rows
pd.options.display.max_rows = None

# check the output of the data aggregation on 100 rows - Clear current outputs afterwards in Cell menu
ords_prods_merge[['user_id', 'order_number', 'max_order']].head(100)

#05 Create 'loyalty_flag' column

# Next Step - loyal customer criteria
ords_prods_merge.loc[ords_prods_merge['max_order'] > 40, 'loyalty_flag'] = 'Loyal customer'

ords_prods_merge.loc[(ords_prods_merge['max_order'] <= 40) & (ords_prods_merge['max_order'] > 10), 'loyalty_flag'] = 'Regular customer'

ords_prods_merge.loc[ords_prods_merge['max_order'] <= 10, 'loyalty_flag'] = 'New customer'

# print the frquency of 'loyalty_flag' with value_counts()
ords_prods_merge['loyalty_flag'].value_counts(dropna = False)

#check the results of just the relevant columns
ords_prods_merge[['user_id', 'loyalty_flag', 'order_number']].head(60)

#06 Analyze spending habits of customers

# Determine the mean price grouped by loyalty_flag
ords_prods_merge.groupby('loyalty_flag').agg({'prices': ['mean']})

At first glance, there is only a small difference in the average price per loyalty category. Other avenues for examination might look for total number of items purchased or total order amount. 

#07 Create a spending flag based on average price by user

# First step - create a new column called 'mean_spending', then group ords_prods_merge by 'user_id', and transform() function 
# applied on the 'prices' column with np.mean arguement
ords_prods_merge['mean_spending'] = ords_prods_merge.groupby(['user_id'])['prices'].transform(np.mean)

#check the new column with .head()
ords_prods_merge.head(15)

# check the output of the data aggregation on 100 rows - Clear current outputs afterwards in Cell menu
ords_prods_merge[['user_id', 'prices', 'mean_spending']].head(100)

# check the 'prices' column, because all visible entries are 9
ords_prods_merge['prices'].mean()

# 2nd Step - "Low Spender" criteria
ords_prods_merge.loc[ords_prods_merge['mean_spending'] < 10, 'spending_flag'] = 'Low spender'

# "High Spender" criteria
ords_prods_merge.loc[(ords_prods_merge['mean_spending'] >= 10), 'spending_flag'] = 'High spender'

#check the new column with .head()
ords_prods_merge.head(10)

#check the results of just the relevant columns
ords_prods_merge[['user_id', 'spending_flag', 'mean_spending']].head(10)

# print the frquency of 'spending_flag' with value_counts()
ords_prods_merge['spending_flag'].value_counts(dropna = False)

#08 Create a frequent customer flag

# First step - create a new column called 'median_order_flag', then group ords_prods_merge by 'user_id', and transform() function 
# applied on the 'days_since_prior_order' column with np.median arguement
ords_prods_merge['median_order_freq'] = ords_prods_merge.groupby(['user_id'])['days_since_prior_order'].transform(np.median)

#check the new column with .head()
ords_prods_merge.head()

#check the results of just the relevant columns - Clear current outputs afterwards in Cell menu
ords_prods_merge[['user_id', 'days_since_prior_order', 'median_order_freq']].head(100)

# 2nd Step - "Non-frequent customer" criteria
ords_prods_merge.loc[(ords_prods_merge['median_order_freq'] > 20) , 'order_freq_flag'] = 'Non-frequent customer'

# "Regular customer" criteria
ords_prods_merge.loc[(ords_prods_merge['median_order_freq'] > 10) & (ords_prods_merge['median_order_freq'] <= 20) , 'order_freq_flag'] = 'Regular customer'

# "Frequent customer" criteria
ords_prods_merge.loc[(ords_prods_merge['median_order_freq'] <= 10) , 'order_freq_flag'] = 'Frequent customer'

# print the frquency of 'order_freq_flag' with value_counts()
ords_prods_merge['order_freq_flag'].value_counts(dropna = False)

#09 Export Data

#check the imported data
ords_prods_merge.shape

ords_prods_merge.head()

#Export df_orders_products_merged in .pkl format
ords_prods_merge.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merge_withflags.pkl'))
