==> 03 Scripts/9_1_Customer_Dataset_Checks_Merging.ipynb <==
# 9.1 Customer Data set, data checks & merging

Contents

1. Importing libraries
2. Importing Data
3. Examine the customer dataframe
4. Data Wrangling
    4.1 Renaming columns
    4.2 Reducing data types 
5. Data Consistency Checks  
6. Merge Customer data with df_orders
7. Export data

#01 Importing libraries

#Import libraries
import pandas as pd
import numpy as np
import os

#02 Importing Data

#Folder Path
path = r'C:\Users\Caroline\Documents\Career Foundry\03 2023 Instacart Basket Analysis'

#import customers dataset as df
df = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'customers.csv'))

#import orders_products_merged.pkl as df_orders
df_orders = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merge_withflags.pkl'))

#check df_orders after importing
df_orders.shape

#03 Examine the customer dataframe

# determine intial number of rows & columns of dataframe
df.shape

# Print the first 5 rows of your dataframe.
df.head()

# Determine the initial descriptive statistic of dataframe
df.info()

df.describe()

#04 Data Wrangling

Dropping columns - The First Name and Surnam columns are not relevant for data analysis and can be dropped. 'user_id' will provide a unique identifier. 

# drop First Name and Surnam columns
df_dropped = df.drop(columns = ['First Name' , 'Surnam'])

#04.1 Renaming columns

#rename 'STATE' to 'state'  'fam_status' to 'family_status'
df_dropped.rename(columns = {'STATE' : 'state' }, inplace = True)

#rename 'Gender' to 'gender'
df_dropped.rename(columns = {'Gender' : 'gender'}, inplace = True)

#rename 'Age' to 'age'
df_dropped.rename(columns = {'Age' : 'age'}, inplace = True)

#rename 'n_dependants' to 'number_dependants'
df_dropped.rename(columns = {'n_dependants' : 'number_dependants'}, inplace = True)

#rename 'fam_status' to 'family_status'
df_dropped.rename(columns = {'fam_status' : 'family_status'}, inplace = True) 

#check the dataframe
df_dropped.info()

#04.2 Reducing data types 

df_dropped.describe()

#change the data type of 'age' and 'number_dependants' to int8
df_dropped['age'] = df_dropped['age'].astype('int8')
df_dropped['number_dependants'] = df_dropped['number_dependants'].astype('int8')

#change the data type of 'user_id' to string and 'income' to int32
df_dropped['user_id'] = df_dropped['user_id'].astype('str')
df_dropped['income'] = df_dropped['income'].astype('int32')

#check the dataframe 
df_dropped.info()

#05 Data Consistency Checks

#examine the dataframe for accuracy
df_dropped.describe()

Minimum, mean, and maximum values all look reasonable. 

# to find missing values 
df_dropped.isnull().sum()

# check for full duplicates
df_dups = df_dropped[df_dropped.duplicated()]

df_dups

No full duplicates in the dataframe, contine to use df_dropped

# Search for mixed data types in df_dropped
for col in df_dropped.columns.tolist():
  weird = (df_dropped[[col]].applymap(type) != df_dropped[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_dropped[weird]) > 0:
    print (col)
print ('test complete - this statement will always print')

Customer dataframe has no missing values, no duplicates, and no mixed data types.

#06 Merge Customer data with df_orders

#check the df_orders for user_id data type 
df_orders.info()

# Merge df_orders and df_dropped. Indicator = true to check if data is present in both dataframes
df_merged = df_orders.merge(df_dropped, on = 'user_id', indicator = True)

#Check output of merged dataframes
df_merged.head()

# Strip the column(s) you're planning to join with
df_orders['user_id'] = df_orders['user_id'].str.strip()

# Strip the column(s) you're planning to join with
df_dropped['user_id'] = df_dropped['user_id'].str.strip()

# Merge df_orders and df_dropped. Indicator = true to check if data is present in both dataframes
df_merged = df_orders.merge(df_dropped, on = 'user_id', indicator = True)

#Check output of merged dataframes
df_merged.head()

# Check value count for '_merge' column
df_merged['_merge'].value_counts()

# drop '_merge' column
df_merged = df_merged.drop(columns = ['_merge'])

#Check df_merged_large for file size
df_merged.info()

#07 Export data

#Export df_merged in .pkl format
df_merged.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'instacart_data_all.pkl'))
